+++ R/mock-session.R
@@ -31,7 +31,6 @@ extract <- function(promise) {
 #' @noRd
 #' @export
 `$.mockclientdata` <- function(x, name) {
-  if (name == "allowDataUriScheme") { return(TRUE) }
   if (name == "pixelratio") { return(1) }
   if (name == "url_protocol") { return("http:") }
   if (name == "url_hostname") { return("mocksession") }
@@ -167,12 +166,10 @@ makeExtraMethods <- function() {
   ), makeErrors(
     `@uploadEnd` = "for internal use only",
     `@uploadInit` = "for internal use only",
-    `@uploadieFinish` = "for internal use only",
     createBookmarkObservers = "for internal use only",
     dispatch = "for internal use only",
     handleRequest = "for internal use only",
     requestFlush = "for internal use only",
-    saveFileUrl = "for internal use only",
     startTiming = "for internal use only",
     wsClosed = "for internal use only"
   ))
@@ -249,6 +246,8 @@ MockShinySession <- R6Class(
     #' @field user The username of an authenticated user. Always `NULL` for a
     #'   `MockShinySession`.
     user = NULL,
+    #' @field options A list containing session-level shinyOptions.
+    options = NULL,
 
     #' @description Create a new MockShinySession.
     initialize = function() {
@@ -273,6 +272,10 @@ MockShinySession <- R6Class(
       self$input <- .createReactiveValues(private$.input, readonly = TRUE)
 
       self$token <- createUniqueId(16)
+
+      # Copy app-level options
+      self$options <- getCurrentAppState()$options
+
       self$cache <- MemoryCache$new()
       self$appcache <- MemoryCache$new()